+++ xenon-common/src/main/java/com/vmware/xenon/services/common/LuceneDocumentIndexService.java
@@ -1824,7 +1824,7 @@ public class LuceneDocumentIndexService extends StatelessService {
         if (queryOptions.contains(QueryOption.INDEXED_METADATA)) {
             if (!queryOptions.contains(QueryOption.INCLUDE_ALL_VERSIONS)
                     && !queryOptions.contains(QueryOption.TIME_SNAPSHOT)) {
-                Query currentClause = NumericDocValuesField.newExactQuery(
+                Query currentClause = NumericDocValuesField.newSlowExactQuery(
                         LuceneIndexDocumentHelper.FIELD_NAME_INDEXING_METADATA_VALUE_TOMBSTONE_TIME,
                         LuceneIndexDocumentHelper.ACTIVE_DOCUMENT_TOMBSTONE_TIME);
                 builder.add(currentClause, Occur.MUST);
@@ -1833,7 +1833,7 @@ public class LuceneDocumentIndexService extends StatelessService {
             // apply the optimization to limit the resultset only when there is no sort specified
             if ((qs != null && qs.sortTerm == null) &&
                     queryOptions.contains(QueryOption.TIME_SNAPSHOT)) {
-                Query tombstoneClause = NumericDocValuesField.newRangeQuery(
+                Query tombstoneClause = NumericDocValuesField.newSlowRangeQuery(
                         LuceneIndexDocumentHelper.FIELD_NAME_INDEXING_METADATA_VALUE_TOMBSTONE_TIME,
                         qs.timeSnapshotBoundaryMicros, LuceneIndexDocumentHelper.ACTIVE_DOCUMENT_TOMBSTONE_TIME);
                 builder.add(tombstoneClause, Occur.MUST);
@@ -1978,8 +1978,7 @@ public class LuceneDocumentIndexService extends StatelessService {
 
         if (qs.groupResultLimit != null && groups.groups.length >= groupLimit) {
             // check if we need to generate a next page for the next set of group results
-            groups = groupingSearch.search(s, tq, groupLimit + groupOffset, groupLimit);
-            if (groups.totalGroupedHitCount > 0) {
+            if (groups.totalHitCount > groups.totalGroupedHitCount) {
                 rsp.nextPageLink = createNextPage(op, s, qs, tq, sort,
                         null, 0, groupLimit + groupOffset,
                         task.documentExpirationTimeMicros, task.indexLink, task.nodeSelectorLink, page != null);
@@ -3538,7 +3537,7 @@ public class LuceneDocumentIndexService extends StatelessService {
 
         Query selfLinkClause = new TermQuery(new Term(ServiceDocument.FIELD_NAME_SELF_LINK,
                 info.selfLink));
-        Query currentClause = NumericDocValuesField.newExactQuery(
+        Query currentClause = NumericDocValuesField.newSlowExactQuery(
                 LuceneIndexDocumentHelper.FIELD_NAME_INDEXING_METADATA_VALUE_TOMBSTONE_TIME,
                 LuceneIndexDocumentHelper.ACTIVE_DOCUMENT_TOMBSTONE_TIME);