+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestLuceneDocumentIndexService.java
@@ -314,7 +314,7 @@ public class TestLuceneDocumentIndexService {
     public Integer expirationSeconds = null;
 
     private final String EXAMPLES_BODIES_FILE = "example_bodies.json";
-    private final String INDEX_DIR_NAME = "lucene510";
+    private final String INDEX_DIR_NAME = "lucene650";
 
     private FaultInjectionLuceneDocumentIndexService indexService;